# apps/api/src/main/java/com/finrhythm/api/diagnostic/web/DiagnosticRoutingAnswerResponse.java
package com.finrhythm.api.diagnostic.web;

import io.swagger.v3.oas.annotations.media.Schema;

@Schema(description = "Q1-Q3 routing-preview answer response.")
public record DiagnosticRoutingAnswerResponse(
        @Schema(description = "Routing question id.", example = "Q1")
        String id,
        @Schema(description = "Selected option id.", example = "B")
        String optionId
) {
}